cwl/conda_requirements.txt
bioconductor-rcwl=1.18.*
python=3.10.*
pyyaml=5.4.*
r-base=4.3.*
